+++ CHANGELOG.md
@@ -25,3 +25,4 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Implement mathematica programs
 - Add labels for main results
 - Update files structure
+- Add forward and backward recurrences for Tm